+++ src/tropology/api.clj
@@ -86,19 +86,22 @@
 
 (defn tropes-from-node
   [code]
-  (let [html   (ut/if-empty (db/get-html code) "")
+  (let [to-get (if (= code "/") (db/fetch-random-contents-code) code)
+        node   (db/query-by-code to-get)
+        html   (-> (if (:is-redirect node) (:redirects-to node) to-get)
+                   db/get-html
+                   (ut/if-empty ""))
         res    (-> html java.io.StringReader. e/html-resource)
         tropes (p/get-tropes res)
         links  (map p/process-links tropes)
         node   (p/node-data-from-meta res)
         ]
-
     (if (empty? html)
       nil
       {:title       (:title node)
+       :image       (:image node)
        :description (:description node)
-       :code        code
+       :code        (:code node)
        :display     (:display node)
-       :tropes      (map :hiccup links)})
-    )
+       :tropes      (map :hiccup links)}))
   )